Fix CI compilation errors and warnings
- Fixed variable redefinition and block-scope declaration in test_codegen_client_body.h
- Fixed sign-compare warning in client_gen.c by explicitly casting to size_t
- Fixed unused parameter warnings in db_loader.c using (void) casting
- Corrected FetchContent_Declare URLs to point to SamuelMarks repository with GIT_TAG master
- Maintained C89 compatibility and integer error checking semantics

+++ src/routes/emit/client_gen.c
@@ -3706,7 +3706,8 @@ int openapi_client_generate(const struct OpenAPI_Spec *spec,
 
           in = spec->paths[i].route;
           out = formatted_path;
-          while (*in && (out - formatted_path) < sizeof(formatted_path) - 2) {
+          while (*in &&
+                 (size_t)(out - formatted_path) < sizeof(formatted_path) - 2) {
             if (*in == '{') {
               *out++ = '1';
               while (*in && *in != '}')
